+++ Xtzkt.Indexers.Common/Cache/XChainCache.cs
@@ -86,7 +86,7 @@ public class XChainCache(XtzktContext db, IConfiguration config) : IChainCache
         return parent.Id | (long)parent.SubsCounter.Value;
     }
 
-    public void ReleaseOperationId(int count = 1)
+    public void ReleaseOperationId(long count = 1)
     {
         Chain.OperationCounter -= count;
     }
